+++ tests/features/test_get_features.py
@@ -825,10 +825,6 @@ def test_partition_asymmetry():
                                                                              0.5,  0.]))
 
 
-def test_partition_asymmetry_length():
-    assert_allclose(get_feature('partition_asymmetry_length', POP)[:1], np.array([0.853925]))
-
-
 def test_section_strahler_orders():
     path = Path(SWC_PATH, 'strahler.swc')
     n = nm.load_neuron(path)